Extracted ProgramDetails, SetRepFields, ExerciseFields
