+++ frontend/src/components/pages/planner/index.ts
@@ -1,4 +1,5 @@
 import { Planner } from './Planner'
 import { ProgramList } from './ProgramList'
+import { WorkoutList } from './WorkoutList'
 
-export { Planner, ProgramList }
+export { Planner, ProgramList, WorkoutList }
