+++ frontend/src/components/pages/workoutDetails/index.ts
@@ -1,0 +1,4 @@
+import { WorkoutDetails } from './WorkoutDetails'
+import { WorkoutDetailsForm } from './WorkoutDetailsForm'
+
+export { WorkoutDetails, WorkoutDetailsForm }
