+++ frontend/src/components/pages/workoutHistory/WorkoutHistoryTableRow.tsx
