+++ frontend/src/components/pages/workoutHistory/index.ts
@@ -1,0 +1,4 @@
+import { WorkoutHistory } from './WorkoutHistory'
+import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
+
+export { WorkoutHistory, WorkoutHistoryTableRow }
