--- frontend/src/components/program/components/index.ts
@@ -1,3 +1,0 @@
-import { ProgramDetails } from './ProgramDetails'
-
-export { ProgramDetails }
