--- frontend/src/components/workout/components/index.ts
@@ -1,13 +1,0 @@
-import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
-import { WorkoutDetails } from './WorkoutDetails'
-import { WorkoutList } from './WorkoutList'
-import { WorkoutHistory } from './WorkoutHistory'
-import { WorkoutDetailsForm } from './WorkoutDetailsForm'
-
-export {
-  WorkoutDetails,
-  WorkoutHistoryTableRow,
-  WorkoutList,
-  WorkoutHistory,
-  WorkoutDetailsForm
-}
